Rename _displayManager to _displayController and split test file
Rename the stale _displayManager variable to _displayController in
FanatecWheelDeviceInstance and ScreenSettingsPanel to match the class
name established in the LayerStackEvaluator extraction.

Split the monolithic DisplayManagerTests.cs into focused test files
matching the one-class-per-file convention used elsewhere:
- SegmentRenderingTests
- DisplaySettingsTests
- LayerCatalogTests
- DisplayLayerTests
- LayerStackEvaluatorTests
